One small bugfix - don't try to calculate insert size unless paired-end alignment is concordant.  Also, improvements to test infrastructure so that error messages are more likely to get passed to the top.

+++ aligner_sw.cpp
@@ -817,7 +817,7 @@ inline void SwAligner::updateNucDiag(
 				return;
 			}
 			COUNT_N(dcOallBest);
-			AlnScore ex = dcOallBest + add;
+			AlnScore ex = dcOallBest + (int)add;
 			if(ex.score_ >= floorsc_ && ex >= myOallBest) {
 				if(ex > myOallBest) {
 					myOallBest = ex;
